Now headless paths will be scanned only once each.
Previous algorithm would scan the same subpath under each head, which can
be very costly when there are many files in each head (cost dominated by
lstat call count). This is because that algorithm was not checking for
identical file subsets inside a same subpath, but only a leading identical
sequence.
With this change, all file subsets for a given subpath are identified in
a single pass.
This accelerates all heads after the first one (for the common backup case
where the majority of files are common between heads - like yesterday's
Maildir compared to today's).
The flip side is that memory usage becomes proportional to the number of
headless paths and their length. Should this become an issue, an
improvement could be to store hashes instead.